+++ travis/drafter_installation.sh
@@ -1,0 +1,5 @@
+git clone --recursive --branch v0.1.9 --depth 1 git://github.com/apiaryio/drafter.git
+cd drafter
+./configure
+make drafter
+sudo make install